+++ pom.xml
@@ -8,7 +8,7 @@
 		<version>2.7.0-M1</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>hh.slackbot</groupId>
+	<groupId>hh.nitor.slackbot</groupId>
 	<artifactId>slackbot</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>slackbot</name>
