+++ src/main/java/hh/nitor/slackbot/BlockActionHandler.java
@@ -1,4 +1,4 @@
-package hh.slackbot.slackbot;
+package hh.nitor.slackbot;
 
 import com.google.gson.JsonObject;
 import com.slack.api.app_backend.interactive_components.payload.BlockActionPayload;
@@ -7,14 +7,14 @@ import com.slack.api.bolt.context.builtin.ActionContext;
 import com.slack.api.bolt.request.builtin.BlockActionRequest;
 import com.slack.api.bolt.response.Response;
 import com.slack.api.model.Usergroup;
-import hh.slackbot.slackbot.util.MessageUtil;
-import hh.slackbot.slackbot.util.RestService;
-import hh.slackbot.slackbot.util.UsergroupUtil;
 import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import hh.nitor.slackbot.util.MessageUtil;
+import hh.nitor.slackbot.util.RestService;
+import hh.nitor.slackbot.util.UsergroupUtil;
 
 @Component
 public class BlockActionHandler {
