+++ src/test/java/hh/nitor/slackbot/BlockActionTests.java
@@ -1,4 +1,4 @@
-package hh.slackbot.slackbot;
+package hh.nitor.slackbot;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
@@ -16,9 +16,9 @@ import com.slack.api.app_backend.interactive_components.payload.BlockActionPaylo
 import com.slack.api.bolt.context.builtin.ActionContext;
 import com.slack.api.bolt.request.builtin.BlockActionRequest;
 import com.slack.api.model.Usergroup;
-import hh.slackbot.slackbot.util.MessageUtil;
-import hh.slackbot.slackbot.util.RestService;
-import hh.slackbot.slackbot.util.UsergroupUtil;
+import hh.nitor.slackbot.util.MessageUtil;
+import hh.nitor.slackbot.util.RestService;
+import hh.nitor.slackbot.util.UsergroupUtil;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
