+++ src/test/java/hh/nitor/slackbot/UsergroupUtilTests.java
@@ -1,4 +1,4 @@
-package hh.slackbot.slackbot;
+package hh.nitor.slackbot;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
@@ -8,7 +8,7 @@ import com.slack.api.methods.SlackApiException;
 import com.slack.api.methods.request.usergroups.UsergroupsListRequest;
 import com.slack.api.methods.response.usergroups.UsergroupsListResponse;
 import com.slack.api.model.Usergroup;
-import hh.slackbot.slackbot.util.UsergroupUtil;
+import hh.nitor.slackbot.util.UsergroupUtil;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
